reviewer 1
I am not sure if this applies differently on a Mac if that's what you used, but I used an android emulator and when I clicked on any of your main 3 tabs it came up with the error "Unfortunately Bus 353 assn. 4 closed. The reason it does this is that where you have the code to open the window, you included code to change the other pages. if you take the code to close the other pages out, then the window will open properly. Other than that, it was a very good use of space and good information in the app. The only other critique I would give is to resize the return button. "retur" is on one level and the "r" is on the next level. Overall, when the parts of the code that aren't needed is taken out, it is a very good app.